extract inner class
