+++ test-harness/src/main/java/edu/umd/cs/findbugs/AbstractIntegrationTest.java
@@ -25,8 +25,6 @@ import static org.junit.Assert.fail;
 import java.io.File;
 import java.io.IOException;
 
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 
 import edu.umd.cs.findbugs.config.UserPreferences;
@@ -143,41 +141,6 @@ public class AbstractIntegrationTest {
                 .map(error -> error.getCause())
                 .forEach(assertionError::addSuppressed);
             throw assertionError;
-        }
-    }
-
-    private static final class CountMatcher<T> extends BaseMatcher<Iterable<T>> {
-        private final int count;
-        private final Matcher<T> matcher;
-
-        /**
-         * @param count
-         * @param matcher
-         */
-        private CountMatcher(int count, Matcher<T> matcher) {
-            this.count = count;
-            this.matcher = matcher;
-        }
-
-        @Override
-        public boolean matches(final Object obj) {
-            int matches = 0;
-
-            if (obj instanceof Iterable<?>) {
-                final Iterable<?> it = (Iterable<?>) obj;
-                for (final Object o : it) {
-                    if (matcher.matches(o)) {
-                        matches++;
-                    }
-                }
-            }
-
-            return matches == count;
-        }
-
-        @Override
-        public void describeTo(final Description desc) {
-            desc.appendText("Iterable containing exactly " + count + " ").appendDescriptionOf(matcher);
         }
     }
 }
